+++ CHANGELOG.md
@@ -2,15 +2,22 @@
 All notable changes to this project will be documented in this file.
 
 #### 1.x Releases
-- `2.6.x` Releases  - [2.6.0](#260) | [2.6.1](#261) | [2.6.2](#262)
+- `2.6.x` Releases  - [2.6.0](#260) | [2.6.1](#261) | [2.6.2](#262) | [2.6.2](#262)
 - `2.5.x` Releases  - [2.5.0](#250)
 
 ---
+## [2.6.3](https://github.com/congncif/IDMFoundation/releases/tag/2.6.3)
+Released on 2018-4-19
+
+#### Updated
+- `IntegrationCall` nextSuccess, nextError : configuration with parameters
+- `IntegrationCall`  added transformNext
+
 ## [2.6.2](https://github.com/congncif/IDMFoundation/releases/tag/2.6.2)
 Released on 2018-4-17
 
 #### Updated
-`IntegrationCall` nextSuccess, nextError with parametersBuilder
+- `IntegrationCall` nextSuccess, nextError with parametersBuilder
 
 ## [2.6.1](https://github.com/congncif/IDMFoundation/releases/tag/2.6.1)
 Released on 2018-4-11